api/requirements.txt
python-dotenv
sqlmodel
fastapi
uvicorn